data/requirements.txt
langchain-community
pypdf
faiss-cpu 
transformers 
sentence-transformers 
huggingface-hub 
langchain-google-genai
python-dotenv